added a table page_keyword
