+++ .gitignore
@@ -1,7 +1,12 @@
 
 .DS_Store
+__pycache__/
+*.py[cod]
+*$py.class
+
 
 .idea/
 database/database.db
+database/__pycache__/
